LTR #23: Merged Feature weight and scorer
